Calendar: Screen is covered by a piece of gray when rotate
When press "next" in IME OpenWnn, Inputmethod Manager can
not be active, and lost its servedView cause it can not
hide normally.

Add android configChanges tag for EditEventActivity, so
it would not to create new instance of EditEventActivity
when rotate the phone, also can avoid showing blank IME
UI when its in deactive state.

CRs-Fixed: 442057

Change-Id: Ib4db5f31fa69b6f00b50a913f9c92eeea5b1b545